+++ vitamin-k/window/vikWindowXCB.hpp
@@ -97,7 +97,7 @@ class WindowXCB : public Window {
     if (settings->fullscreen) {
       get_randr_outputs();
 
-      if (settings->display > displays.size() - 1) {
+      if (settings->display > (int) displays.size() - 1) {
         vik_log_w("Requested display %d, but only %d displays are available.",
                   settings->display, displays.size());
